Add theme provider using matchMedia
