+++ src/speedscope.tsx
@@ -2,9 +2,7 @@ import {h, render} from 'preact'
 import {createAppStore} from './store'
 import {ApplicationContainer} from './views/application-container'
 import {Provider} from './lib/preact-redux'
-import { ThemeContext } from './views/themes/theme'
-import { darkTheme } from './views/themes/dark-theme'
-import { lightTheme } from './views/themes/light-theme'
+import { ThemeProvider } from './views/themes/theme'
 
 console.log(`speedscope v${require('../package.json').version}`)
 
@@ -22,11 +20,11 @@ const store = lastStore ? createAppStore(lastStore.getState()) : createAppStore(
 ;(window as any)['store'] = store
 
 render(
-  <Provider store={store}>
-    <ThemeContext.Provider value={darkTheme}>
-      <ApplicationContainer />
-    </ThemeContext.Provider>
-  </Provider>,
+  <ThemeProvider>
+    <Provider store={store}>
+        <ApplicationContainer />
+    </Provider>
+  </ThemeProvider>,
   document.body,
   document.body.lastElementChild || undefined,
 )
