--- src/views/themes/theme.ts
@@ -1,38 +1,0 @@
-import {createContext} from 'preact'
-import {useContext} from 'preact/hooks'
-import {Color} from '../../lib/color'
-import {memoizeByReference} from '../../lib/utils'
-import {lightTheme} from './light-theme'
-
-export interface Theme {
-  fgPrimaryColor: string
-  fgSecondaryColor: string
-  bgPrimaryColor: string
-  bgSecondaryColor: string
-
-  altFgPrimaryColor: string
-  altFgSecondaryColor: string
-  altBgPrimaryColor: string
-  altBgSecondaryColor: string
-
-  selectionPrimaryColor: string
-  selectionSecondaryColor: string
-
-  weightColor: string
-
-  searchMatchPrimaryColor: string
-  searchMatchSecondaryColor: string
-
-  colorForBucket: (t: number) => Color
-  colorForBucketGLSL: string
-}
-
-export const ThemeContext = createContext<Theme>(lightTheme)
-
-export function useTheme(): Theme {
-  return useContext(ThemeContext)
-}
-
-export function withTheme<T>(cb: (theme: Theme) => T) {
-  return memoizeByReference(cb)
-}
